requirements.txt
azure-ai-agents>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
